# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * GEOscore API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuditInput {
  url: string;
  /** @nullable */
  brandName?: string | null;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  market?: string | null;
}

export type RecommendationPriority = typeof RecommendationPriority[keyof typeof RecommendationPriority];


export const RecommendationPriority = {
  high: 'high',
  medium: 'medium',
  low: 'low',
} as const;

export interface Recommendation {
  title: string;
  description: string;
  priority: RecommendationPriority;
  aiSystem: string;
}

export interface AuditResult {
  id: string;
  url: string;
  domain: string;
  /** @nullable */
  brandName?: string | null;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  market?: string | null;
  scoreTotal: number;
  scoreChatgpt: number;
  scoreGemini: number;
  scorePerplexity: number;
  chatgptFound: boolean;
  geminiFound: boolean;
  perplexityFound: boolean;
  /** @nullable */
  chatgptDetail?: string | null;
  /** @nullable */
  geminiDetail?: string | null;
  /** @nullable */
  perplexityDetail?: string | null;
  competitorsFound?: string[];
  keywordsUsed?: string[];
  recommendations?: Recommendation[];
  createdAt: string;
}

export interface AuditSummary {
  id: string;
  domain: string;
  /** @nullable */
  brandName?: string | null;
  scoreTotal: number;
  scoreChatgpt?: number;
  scoreGemini?: number;
  scorePerplexity?: number;
  chatgptFound?: boolean;
  geminiFound?: boolean;
  perplexityFound?: boolean;
  createdAt: string;
}

export interface RegisterInput {
  email: string;
  password: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export type UserPlan = typeof UserPlan[keyof typeof UserPlan];


export const UserPlan = {
  free: 'free',
  starter: 'starter',
  agency: 'agency',
} as const;

export interface User {
  id: string;
  email: string;
  plan: UserPlan;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface MonitoredBrand {
  id: string;
  domain: string;
  /** @nullable */
  brandName?: string | null;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  market?: string | null;
  keywords?: string[];
  competitors?: string[];
  /** @nullable */
  latestScore?: number | null;
  /** @nullable */
  latestScoreChatgpt?: number | null;
  /** @nullable */
  latestScoreGemini?: number | null;
  /** @nullable */
  latestScorePerplexity?: number | null;
  /** @nullable */
  lastChecked?: string | null;
  createdAt: string;
}

export interface MonitoredBrandInput {
  domain: string;
  /** @nullable */
  brandName?: string | null;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  market?: string | null;
  keywords?: string[];
  competitors?: string[];
}

export interface DashboardSummary {
  totalBrands: number;
  avgScore: number;
  brandsWithImprovement: number;
  totalAuditsRun: number;
  /** @nullable */
  topBrand?: string | null;
  /** @nullable */
  topBrandScore?: number | null;
}

export interface DailyScore {
  id: string;
  brandId: string;
  date: string;
  scoreTotal: number;
  scoreChatgpt: number;
  scoreGemini: number;
  scorePerplexity: number;
  recommendations?: Recommendation[];
}

export type PaymentOrderInputPlan = typeof PaymentOrderInputPlan[keyof typeof PaymentOrderInputPlan];


export const PaymentOrderInputPlan = {
  starter: 'starter',
  agency: 'agency',
} as const;

export interface PaymentOrderInput {
  plan: PaymentOrderInputPlan;
}

export interface PaymentOrder {
  orderId: string;
  amount: number;
  currency: string;
  keyId: string;
}

export interface PaymentVerifyInput {
  razorpayOrderId: string;
  razorpayPaymentId: string;
  razorpaySignature: string;
  plan: string;
}

export interface EmailSubscribeInput {
  email: string;
  /** @nullable */
  domain?: string | null;
  /** @nullable */
  auditId?: string | null;
}