# lib/api-zod/src/generated/types/dailyScore.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * GEOscore API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Recommendation } from './recommendation';

export interface DailyScore {
  id: string;
  brandId: string;
  date: string;
  scoreTotal: number;
  scoreChatgpt: number;
  scoreGemini: number;
  scorePerplexity: number;
  recommendations?: Recommendation[];
}